Big Bug fixes
Fixed bug where we would get stuck on sides of platforms by adding frictionless physics material to the platforms then adding a child square object that would rest at the top of the platform allowing friction to occur when the plauer is on it.

Also changed players box collider to a capsule collider which is doing wonders. First of all I dont know why I started with a box collider in the first place when the capsule works so much better fit wise. this also made slopes more predictable and the player can now walk up slopes up to 45 degrees -ish and there isnt much jitter. I had to increase the ground friction because with the capsule there is less touching the ground and thus smaller frictional force so again to compensate I increased the drag back up to 2.0f for now.