QOL changes
Changed collider size to fit head better

changed camera so cant clip look through platforms

added control variables script and shoved it on an empty game object. This will have things like keybinds and mouse sensitivity on it and other things that would potentially be changeable by a player in the options.