Tried Camera connection via script
made camera no longer child of player because heard having a camera on a object that has Rigidbody could cause issues and I think I was seeing a little camera jitter. I switched it and as of now camera jitter is the same so I'm making this branch so if theres an issue in the future then we can merge it back into main and see if it helps then, but for now may just be a useless branch.