Movement States
Added Sprinting and crouching capabilities and made each state modify the max speed the player can move at. Also added a velocity stat on screen so its easier to see while testing. Also made game manager which for now just allows me to work with the canvas off until runtime so I dont have to see the words when navigating the scene. Most of the files are because TextMeshPro stuff needed to import so I could use it properly.