Sprinting Crouching and MAYBE wall sticking better fix
Changed it so the player is frictionless rather than all the ground peices and this seems to work with stopping wall sticking. Added sprinting and the stamina bar and functionality. It depletes when sprinting and after 10s you're out of sprint juice and once you stop holding sprint itll begin to recharge. Then I implemented the crouching. Made local scale smaller and had to change the capsule radius smaller to so we could fit under things. made a can stand check to see if player tries to stand up while under something. Made a little tunnel to see what crouching through feels like. SLOPES NEED TO BE DONE, between changing friction material on the player and the ground friction the player can now climb STEEP slopes. I need to figure out a way to limit it.