Changed crouch, moveStates conditions, and ground Drag
Changed crouching standing detection to work more accurately by only detecting if the thing above it is part of the whatIsGround layer mask. This fixed the issue of what I assume was the sphereCast running into the player object itself and occasionally not allowing the player to stand even when it seemed clear.

Then also fixed the move state changes conditions to ensure there were no issues or bugs the player could get by smashing the buttons.

Changed ground drag from 5 to 8 to lower the level of sliding the player feels when tying to come to a stop naturally. Then also needed to increase jumpForce a tad to compensate for the increased drag that impulse force feels when it makes the character jump.